Fix rendering of allowed characters.
The default rule for sphinx with simple backticks seem to be "cite",
which is rendered in italic, making it relatively hard to distinguish
what is the set of characters meant vs, formatting.

On my initial read on current page I actually though the trailing dot
was finishing the sentence, and it was not clear wether or not spaces
(and or commas) were part of allowed character.

With the double quote each character group is rendered: monospace, and in
a different color (for current theme)