Add the notion of a key suffix to metadata documents
Closes #63

The rational is that backends may be explored by non zarr-aware tools,
like a filesytem and a GUI browser like Finder/Explorer/Nautilus, and
the `.json` extension will help to view/edit json in those.

While it may seem like this shoudl be store specific, once again the
fact that a zarr-hierarchy can be copied from one storage to another by
non-zarr-aware tool indicate that this shoudl be handled above the sotre
level.