Change zarr_format to be just "zarr.dev/v3"
Previously, the specification stated that `zarr_format` should be
"https://purl.org/zarr/spec/core/3.0", with the last component
indicating the major and minor version number, but with the
requirement of forward and backward compatibility across different
minor versions.

However, given the compatibility requirement, it is not clear what
purpose the minor version number serves.

Furthermore, due to the fact that the zarr repository may be created
with one zarr implementation, and later an array or group may be added
using a different zarr implementation/version, it is not clear in
which cases the version number may be updated.

Therefore, this commit removes the minor version, and changes to a
more concise URL in order to make the metadata more easily read and
written by humans.