Revise how the Blosc codec is specified
The `shuffle` mode is now specified more clearly as `null` (0 in Zarr
v2), `"bit"` (2 in Zarr v2), or `"byte"` (1 in Zarr v2).  Using these
constants rather than numbers makes it much easier to know what
shuffle mode will be using from manual inspection of the metadata.

When shuffling is enabled, the `typesize` must now be specified
explicitly in the metadata, rather than determined implicitly from the
input data.  This allows Blosc to function as a pure "bytes -> bytes"
codec rather than an "array -> bytes" codec.

The special `shuffle` value of `-1`, which indicated to choose
automatically between bit-wise or byte-wise shuffling depending on the
typesize value, is not supported in the metadata.