Require array -> bytes codec be specified in metadata
This removes the concept of a "default binary representation".
Instead, it is required that an "array -> bytes" codec be specified in
the list of codecs.  Implementations are still free to choose a
default "array -> bytes" codec when creating an array if not specified
by the user.  This should slightly simplify implementations, and makes
it clearer to humans when reading the metadata which "array -> bytes"
codec will be used.