Require transpose codec `order` option to be specified as an explicit permutation
The "C" and "F" constants were added to emphasize the relationship
with the zarr v2 `order` metadata field.  However, these contants are
somewhat confusing:

- "C" means no-op and therefore never needs to be used

- "F" means reverse dimension order.  If there are no other changes to
the dimesnion order, this does indeed result in Fortran-order storage,
but if there are other changes to the dimension order, that is not the
case.  For example, in the following:

"codecs": [
  {"name": "transpose", {"configuration": {"order": "F"}}},
  {"name": "transpose", {"configuration": {"order": "C"}}},
  {"name": "bytes"}
]

or

"codecs": [
  {"name": "transpose", {"configuration": {"order": "C"}}},
  {"name": "transpose", {"configuration": {"order": "F"}}},
  {"name": "bytes"}
]

the storage order is actually "F", while in the following:

"codecs": [
  {"name": "transpose", {"configuration": {"order": "F"}}},
  {"name": "transpose", {"configuration": {"order": "F"}}},
  {"name": "bytes"}
]

the storage order is actually "C".