Fix Example JSON document for Array metadata.
JSON array metadata document in the **Array metadata** section is
ill-formed. the `codecs` field contains just the gzip codec configuration,
even though the spec requires the codec chain to always have one
`array->bytes` codec. That is in the example the gzip codec should be
replaced with something more sensible like
`{"name": "bytes", "configuration": "big"}` to be considered correct.
This commit fixes that.